fix: refactor rootpage
