+++ lib/constants/theme_constants.dart
@@ -1,0 +1,4 @@
+import 'package:flutter/material.dart';
+
+const primaryColor = Color(0xFF0F1236);	
+const accentColor = Color(0xFF45cab9);
