--- lib/ftu/rootpage.dart
@@ -1,67 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_form_builder/flutter_form_builder.dart';
-import 'package:flutter_redux/flutter_redux.dart';
-import 'package:redux/redux.dart';
-
-import '../actions/actions.dart';
-import '../constants/common_constants.dart';
-import '../models/models.dart';
-import '../navigation/app_controller.dart';
-import '../profile/save_profile/save_profile.dart';
-import '../selectors/selectors.dart';
-import '../services/user_service.dart';
-import './authentication.dart';
-import './get_started/intro.dart';
-
-class RootPage extends StatefulWidget {
-  RootPage({Key key}) : super(key: key);
-
-  @override
-  State<StatefulWidget> createState() => _RootPageState();
-}
-
-class _RootPageState extends State<RootPage>{
-  final GlobalKey<FormBuilderState> _buildProfileKey
-    = GlobalKey<FormBuilderState>();
-
-  @override
-  void initState() {
-    super.initState();
-    getCurrentUser(StoreProvider.of<AppState>(context, listen: false));
-  }
- 
-  void getCurrentUser(Store<AppState> store) async {
-    Auth().getCurrentUser().then((user) async {
-      var userData = await getUser(user?.uid);
-      store.dispatch(ConvertToUserState(userData));
-    });
-  }
-
-  Widget getNextPage(AppState state) {
-    switch (state.userState.authStatus) {
-      case AuthStatus.notDetermined:
-        return loadingScreen;
-      case AuthStatus.notLoggedIn:
-        return GetStartedIntro();
-      case AuthStatus.loggedIn:
-        if (state.userState.isFtu) {
-          return SaveProfile(
-            editMode: false,
-            formKey: _buildProfileKey
-          );
-        }
-        return AppController();
-      default:
-        return loadingScreen;
-    }
-      
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return StoreConnector<AppState, AppState>(
-      converter: appStateSelector,
-      builder: (context, state) => getNextPage(state)
-    );
-  }
-}
